requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
xgboost
tensorflow  
psycopg2
SQLAlchemy
Flask
PyYAML
python-dotenv


